Add update functinality
